Use useSpeechRecognition in example app
