+++ example/src/Dictaphones.js
@@ -3,6 +3,9 @@ import { DictaphoneWidgetA, DictaphoneWidgetB } from './Dictaphone'
 import SpeechRecognition from './SpeechRecognition'
 
 export default () => {
+  const [showFirstWidget, setShowFirstWidget] = useState(true)
+  const toggleShowFirstWidget = () => setShowFirstWidget(!showFirstWidget)
+
   const [transcribing, setTranscribing] = useState(true)
   const toggleTranscribing = () => {
     setTranscribing(!transcribing)
@@ -24,12 +27,13 @@ export default () => {
 
   return (
     <div>
-      <DictaphoneWidgetA />
+      {showFirstWidget && <DictaphoneWidgetA />}
       <DictaphoneWidgetB />
       <button onClick={toggleTranscribing}>Toggle transcribing</button>
       <button onClick={listenOnce}>Listen once</button>
       <button onClick={listenContinuously}>Listen continuously</button>
       <button onClick={listenContinuouslyInChinese}>Listen continuously (Chinese)</button>
+      <button onClick={toggleShowFirstWidget}>Toggle first widget</button>
       <button onClick={SpeechRecognition.stopListening}>Stop</button>
     </div>
   )
